move textv1 in text
