+++ src/palette/elements/EntityHeader/EntityHeader.tsx
@@ -1,8 +1,9 @@
+import { bullet } from "palette/helpers"
 import React from "react"
 import { Avatar } from "../Avatar"
 import { Flex } from "../Flex"
 import { SpacerProps } from "../Spacer"
-import { bullet, Sans } from "../Typography"
+import { Sans } from "../Text"
 
 interface EntityHeaderProps extends SpacerProps {
   smallVariant?: boolean
