+++ src/palette/elements/Text/index.ts
@@ -1,2 +1,4 @@
 export * from "./Text"
 export * from "./tokens"
+
+export { Sans, Sans as TextV1, Serif, _maxWidth, SansProps, SansProps as TextV1Props, SerifProps } from "./Typography" // TextV1
