+++ src/palette/elements/index.tsx
@@ -15,5 +15,4 @@ export * from "./Spacer"
 export * from "./Spinner"
 export * from "./Text"
 export * from "./Touchable"
-export * from "./Typography"
 export * from "./Pill"
