+++ src/palette/helpers/index.ts
@@ -1,1 +1,1 @@
-export * from "./typography"
+export * from "./text"
